Upgrade project to C++20 Upgrade external libs to latest version Convert Assimp to submodule

+++ Engine/Source/Renderer/RHI/D3D12/D3D12Utility.cpp
@@ -4,9 +4,9 @@
 
 uint64_t D3D12Utility::CurrentFrameIndex = 0;
 
-D3D12_RASTERIZER_DESC RasterizerStates[ERasterizerState::Count];
-D3D12_BLEND_DESC BlendStates[EBlendState::Count];
-D3D12_DEPTH_STENCIL_DESC DepthStates[EDepthState::Count];
+D3D12_RASTERIZER_DESC RasterizerStates[uint8_t(ERasterizerState::Count)];
+D3D12_BLEND_DESC BlendStates[uint8_t(EBlendState::Count)];
+D3D12_DEPTH_STENCIL_DESC DepthStates[uint8_t(EDepthState::Count)];
 
 void D3D12Utility::Init()
 {
